+++ src/compas_rhino/objects/networkobject.py
@@ -7,11 +7,13 @@ from compas.geometry import Point
 from compas.geometry import Scale
 from compas.geometry import Translation
 from compas.geometry import Rotation
-from compas_rhino.objects._object import BaseObject
-from compas_rhino.objects.modify import network_update_attributes
-from compas_rhino.objects.modify import network_update_node_attributes
-from compas_rhino.objects.modify import network_update_edge_attributes
-from compas_rhino.objects.modify import network_move_node
+
+from ._modify import network_update_attributes
+from ._modify import network_update_node_attributes
+from ._modify import network_update_edge_attributes
+from ._modify import network_move_node
+
+from ._object import BaseObject
 
 
 __all__ = ['NetworkObject']
@@ -46,10 +48,6 @@ class NetworkObject(BaseObject):
         'show.edgelabels': False,
     }
 
-    modify = network_update_attributes
-    modify_nodes = network_update_node_attributes
-    modify_edges = network_update_edge_attributes
-
     def __init__(self, network, scene=None, name=None, layer=None, visible=True, settings=None):
         super(NetworkObject, self).__init__(network, scene, name, layer, visible)
         self._guid_node = {}
@@ -76,12 +74,6 @@ class NetworkObject(BaseObject):
         self._guid_edge = {}
         self._guid_nodelabel = {}
         self._guid_edgelabel = {}
-
-    # def __getstate__(self):
-    #     pass
-
-    # def __setstate__(self, state):
-    #     pass
 
     @property
     def anchor(self):
@@ -273,6 +265,56 @@ class NetworkObject(BaseObject):
         guids = compas_rhino.select_lines()
         edges = [self.guid_edge[guid] for guid in guids if guid in self.guid_edge]
         return edges
+
+    def modify(self):
+        """Update the attributes of the network.
+
+        Returns
+        -------
+        bool
+            ``True`` if the update was successful.
+            ``False`` otherwise.
+        """
+        return network_update_attributes(self.network)
+
+    def modify_nodes(self, nodes, names=None):
+        """Update the attributes of the nodes.
+
+        Parameters
+        ----------
+        nodes : list
+            The identifiers of the nodes to update.
+        names : list, optional
+            The names of the atrtibutes to update.
+            Default is to update all attributes.
+
+        Returns
+        -------
+        bool
+            ``True`` if the update was successful.
+            ``False`` otherwise.
+
+        """
+        return network_update_node_attributes(self.network, nodes, names=names)
+
+    def modify_edges(self, edges, names=None):
+        """Update the attributes of the edges.
+
+        Parameters
+        ----------
+        edges : list
+            The identifiers of the edges to update.
+        names : list, optional
+            The names of the atrtibutes to update.
+            Default is to update all attributes.
+
+        Returns
+        -------
+        bool
+            ``True`` if the update was successful.
+            ``False`` otherwise.
+        """
+        return network_update_edge_attributes(self.network, edges, names=names)
 
     def move(self):
         """Move the entire mesh object to a different location."""